+++ src/components/Projects/carbon_knight/images/carbon_knight.png
