+++ src/components/Projects/carbon_knight/images/resources.png
