+++ src/components/Projects/pack_smart/images/how_it_works.png
